# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Disciplr API
  version: 0.1.0
  description: API documentation for Disciplr backend
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/jobs/enqueue:
    post:
      summary: Enqueue a background job
      tags:
        - Jobs
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                payload: {}
                delayMs:
                  type: number
                maxAttempts:
                  type: number
              required:
                - type
      responses:
        "202":
          description: Job enqueued
          content:
            application/json:
              schema: {}
...
  /api/jobs/deadletters:
    get:
      summary: List dead-letter jobs
      tags:
        - Jobs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of dead-letter jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  deadLetters:
                    type: array
                    items: {}
                required:
                  - deadLetters
  /api/jobs/deadletters/{id}:
    get:
      summary: Get a dead-letter job by ID
      tags:
        - Jobs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dead-letter job details
        "404":
          description: Job not found
  /api/jobs/deadletters/{id}/replay:
    post:
      summary: Replay a dead-letter job
      tags:
        - Jobs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Job replayed
        "404":
          description: Job not found
...